refactor(drive): refactor drive tests into one single file
- previous TC-8785, TC-8786, TC-8787 and TC-8788 where in separate files, this commit refactors them into one file drive.spec.ts

Refs: WPB-24426
